immediate_sync : Monitor files for data change
This change will create watcher objects for all the files/directories
defined in the phosphor-data-sync JSON config files, monitoring the
IN_CLOSE_WRITE events.

IN_CLOSE_WRITE will be able to notify when file opened for writing
was closed, resulting in catching the data changes of both the
watching files and files inside the watching directories.

The callback function will get trigger upon reception of the events
and will rsync will update the data to the sibling BMC.

Tested:
- Verified the following in SIMICS model

- Modification of watched files generated IN_CLOSE_WRITE and triggered
  rsync wrapper to sync it to the sibling BMC

- Modification/creation of files inside a watched directory also
  generated IN_CLOSE_WRITE and triggered rsync wrapper to sync it to
  the sibling BMC

Change-Id: Ib2753a3f033b253c2ba472b35c7a554a20d9b431
Signed-off-by: Arya K Padman <aryakpadman@gmail.com>